refact: move view models back to domain abstraction
