+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/domain/repository/PokemonRepository.kt
@@ -1,7 +1,7 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository
 
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.models.Pokemon
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.PokemonDataSource
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.models.Pokemon
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.models.PokemonList
 import javax.inject.Inject
