+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/ui/components/PokemonTypeBadge.kt
@@ -3,7 +3,7 @@ package com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.github.felipecastilhos.pokedexandroid.core.ui.components.TextBadge
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.PokemonType
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.PokemonType
 
 @Composable
 fun PokemonTypeBadge(modifier: Modifier = Modifier, type: PokemonType) {
