+++ app/src/test/java/com/github/felipecastilhos/pokedexandroid/PokemonRepositoryTest.kt
@@ -1,7 +1,7 @@
 package com.github.felipecastilhos.pokedexandroid
 
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.models.*
-import com.github.felipecastilhos.pokedexandroid.features.pokemon.data.datasource.repository.DefaultPokemonRemoteDataRepository
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.repository.DefaultPokemonRemoteDataRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
